Version 2.2.0.8 Beta 5 ================= bear: - Various fixes made. - Fixed a problem in the shared memory handler that was causing locking of the file ref count.
git-svn-id: http://svn.nate.deepcreek.org.au/svn/TSFileSource/trunk@125 89fdcf3f-49f1-7541-bbf8-3596d8354d9d

+++ src/DVBMpeg2DataParser.cpp
@@ -694,7 +694,7 @@ HRESULT DVBMpeg2DataParser::ReadSection(DVBSection *pSection)
 	}
 	(log << "Supplied data buffer\n").Write();
 
-	DWORD dwWait = WaitForMultipleObjects(2, m_hScanningStopEvent, FALSE, pSection->timeout*1000);
+	DWORD dwWait = WaitForMultipleObjects(2, m_hScanningStopEvent, FALSE, (DWORD)pSection->timeout*1000);
 	(log << "Event or timeout triggered  " << (long)dwWait << "\n").Write();
 
 	if (dwWait == WAIT_OBJECT_0)
