+++ src/RegStore.cpp
@@ -468,7 +468,7 @@ BOOL CRegStore::removeOld()
 	// Now remove old items
 	//
 	LPTSTR pName = "";
-	for(int x = 0; x < toRemove.size(); x++)
+	for(int x = 0; x < (int)toRemove.size(); x++)
 	{
 		pName = toRemove.at(x);
 		RegDeleteKey(settingsKey, pName);
