+++ src/TSFileSinkPin.cpp
@@ -49,7 +49,7 @@ CTSFileSinkPin::CTSFileSinkPin(CTSFileSink *pTSFileSink, LPUNKNOWN pUnk, CBaseFi
 //Frodo code changes
 
 	m_writeBufferSize = 4096*32;
-	m_writeBuffer = new BYTE[m_writeBufferSize];
+	m_writeBuffer = new BYTE[(UINT)m_writeBufferSize];
     m_writeBufferLen = 0;
 
 	m_WriteSampleSize = 0;
@@ -270,7 +270,7 @@ HRESULT CTSFileSinkPin::Filter(byte* pbData,long sampleLen)
 	for(t=off;t<(DWORD)sampleLen;t+=packet)   
 	{
         //sanity check 
-		if (t+packet > sampleLen)
+		if ((int)t+packet > sampleLen)
 			break;
 		
 		if(!bProg)
@@ -533,13 +533,13 @@ void CTSFileSinkPin::setNumbErrorPackets(__int64 lpllErrors)
 void CTSFileSinkPin::PrintLongLong(LPCTSTR lstring, __int64 value)
 {
 	TCHAR sz[100];
-	double dVal = value;
+	double dVal = (double)value;
 	double len = log10(dVal);
-	int pos = len;
+	int pos = (int)len;
 	sz[pos+1] = '\0';
 	while (pos >= 0)
 	{
-		int val = value % 10;
+		int val = (int)(value % 10);
 		sz[pos] = '0' + val;
 		value /= 10;
 		pos--;
